+++ com/learning/mahesh/part3conditionalstatements/Practice.java
@@ -1,4 +1,4 @@
-package com.learning.mahesh;
+package com.learning.mahesh.part3conditionalstatements;
 
 public class Practice {
     public static void main(String[] args) {
